# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Attendance System API",
        "description": "API Documentation for Attendance System",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/attendance/clock-in": {
            "post": {
                "tags": [
                    "Attendance"
                ],
                "summary": "Clock in an employee",
                "operationId": "7fa3546757b447c76dcd375751727497",
                "requestBody": {
                    "description": "Provide the employee ID to clock in",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "employee_id"
                                ],
                                "properties": {
                                    "employee_id": {
                                        "type": "integer",
                                        "format": "int64",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Clock in successful"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "400": {
                        "description": "Employee already clocked in today"
                    }
                }
            }
        },
        "/api/v1/attendance/clock-out": {
            "put": {
                "tags": [
                    "Attendance"
                ],
                "summary": "Clock out an employee",
                "operationId": "697dbcb480193498e7f9f2fe5da9b941",
                "requestBody": {
                    "description": "Provide the employee ID to clock out",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "employee_id"
                                ],
                                "properties": {
                                    "employee_id": {
                                        "type": "integer",
                                        "format": "int64",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Clock out successful"
                    },
                    "404": {
                        "description": "No active attendance found to clock out"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/v1/attendance-log": {
            "get": {
                "tags": [
                    "Attendance"
                ],
                "summary": "Get attendance log with filters",
                "operationId": "12e18f02df4efb84397d37e5202c27c3",
                "parameters": [
                    {
                        "name": "date",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2025-08-15"
                        }
                    },
                    {
                        "name": "department_id",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/v1/departments": {
            "get": {
                "tags": [
                    "Departments"
                ],
                "summary": "Get list of departments",
                "operationId": "149f8ee66a43550287b8c706fccbb9fd",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Departments"
                ],
                "summary": "Create a new department",
                "operationId": "c7acdb28cde544513edb45929eec541b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "max_clock_in_time",
                                    "max_clock_out_time"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Engineering"
                                    },
                                    "max_clock_in_time": {
                                        "type": "string",
                                        "format": "time",
                                        "example": "09:00"
                                    },
                                    "max_clock_out_time": {
                                        "type": "string",
                                        "format": "time",
                                        "example": "17:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Department created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/v1/departments/{id}": {
            "get": {
                "tags": [
                    "Departments"
                ],
                "summary": "Get a single department",
                "operationId": "213b08ec70ae719037e76b87c6c17d76",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Department not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Departments"
                ],
                "summary": "Update a department",
                "operationId": "449226a8438ae73d48c2d487c005a223",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "max_clock_in_time",
                                    "max_clock_out_time"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Engineering"
                                    },
                                    "max_clock_in_time": {
                                        "type": "string",
                                        "format": "time",
                                        "example": "09:00"
                                    },
                                    "max_clock_out_time": {
                                        "type": "string",
                                        "format": "time",
                                        "example": "17:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Department updated successfully"
                    },
                    "404": {
                        "description": "Department not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Departments"
                ],
                "summary": "Delete a department",
                "operationId": "6b5ef138dcefb819cbbcc160d33c236d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "description": "Department not found"
                    }
                }
            }
        },
        "/api/v1/employees": {
            "get": {
                "tags": [
                    "Employees"
                ],
                "summary": "Get list of employees",
                "operationId": "326f2e86fa40a9ba88b1cb5f2d4b62b8",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Employees"
                ],
                "summary": "Create a new employee",
                "operationId": "4a10b3ada27736863f2586c56b9c19ec",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "position",
                                    "department_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john.doe@example.com"
                                    },
                                    "position": {
                                        "type": "string",
                                        "example": "Software Engineer"
                                    },
                                    "department_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Employee created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/v1/employees/{id}": {
            "get": {
                "tags": [
                    "Employees"
                ],
                "summary": "Get a single employee",
                "operationId": "b7cb6100cedcac60d8c23d3f9d7c3251",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Employee not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Employees"
                ],
                "summary": "Update an employee",
                "operationId": "bcdffcad24dd066f489c0e6ed83861a5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "position",
                                    "department_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john.doe@example.com"
                                    },
                                    "position": {
                                        "type": "string",
                                        "example": "Software Engineer"
                                    },
                                    "department_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Employee updated successfully"
                    },
                    "404": {
                        "description": "Employee not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Employees"
                ],
                "summary": "Delete an employee",
                "operationId": "43aa0966dc246d0b0373a7db65a6bf78",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "description": "Employee not found"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Attendance",
            "description": "API Endpoints for Employee Attendance"
        },
        {
            "name": "Departments",
            "description": "API Endpoints for Departments"
        },
        {
            "name": "Employees",
            "description": "API Endpoints for Employees"
        }
    ]
}